# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/backstage/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User View",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "int default",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            15,
                            20,
                            30,
                            40,
                            50
                        ],
                        "type": "integer",
                        "description": "int enums",
                        "name": "pageLimit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "login_name"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sortColumn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "searchCategory",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserListDTO"
                        }
                    }
                }
            }
        },
        "/backstage/user/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User Create",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/user/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/user/edit/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User Edit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/user/password/edit/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User Passowrd Edit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserEditPwdDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/user/password/reset/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User Passowrd Reset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserEditPwdDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-User"
                ],
                "summary": "User By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.UserIdDTO"
                        }
                    }
                }
            }
        },
...
        "/production/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forestage"
                ],
                "summary": "Production list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "int default",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            15,
                            20,
                            30,
                            40,
                            50
                        ],
                        "type": "integer",
                        "description": "int enums",
                        "name": "pageLimit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "PName",
                            "PId",
                            "PCategory",
                            "PCreTime"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sortColumn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "searchCategory",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forestagedto.ProductionDTO"
                        }
                    }
                }
            }
        },
...
        "/production/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forestage"
                ],
                "summary": "Production detail",
                "operationId": "1",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forestagedto.ProductionDetailDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "backstagedto.UserCreateOrEditDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loginName": {
                    "description": "登入帳號",
                    "type": "string",
                    "minLength": 4
                },
                "name": {
                    "description": "使用者名稱",
                    "type": "string",
                    "minLength": 4
                },
                "password": {
                    "description": "密碼",
                    "type": "string",
                    "minLength": 6
                },
                "remark": {
                    "description": "備註",
                    "type": "string"
                },
                "select": {
                    "description": "選的role",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "帳號狀態(false停用 true正常)",
                    "type": "boolean"
                },
                "userType": {
                    "description": "是否為系統用戶",
                    "type": "boolean"
                }
            }
        },
        "backstagedto.UserEditPwdDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "newPassword": {
                    "description": "新密碼",
                    "type": "string",
                    "minLength": 6
                },
                "orgPassword": {
                    "description": "原始密碼",
                    "type": "string"
                },
                "type": {
                    "description": "1.重置密碼 2.修改密碼",
                    "type": "integer"
                }
            }
        },
        "backstagedto.UserIdDTO": {
            "type": "object",
            "properties": {
                "userById": {
                    "$ref": "#/definitions/backstagedto.UserCreateOrEditDTO"
                }
            }
        },
        "backstagedto.UserListDTO": {
            "type": "object",
            "properties": {
                "pageData": {
                    "$ref": "#/definitions/dto.PageForMultSearchDTO"
                },
                "userList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.UserViewData"
                    }
                }
            }
        },
        "backstagedto.UserViewData": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loginName": {
                    "description": "登入帳號",
                    "type": "string",
                    "minLength": 4
                },
                "name": {
                    "description": "使用者名稱",
                    "type": "string",
                    "minLength": 4
                },
                "remark": {
                    "description": "備註",
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "description": "帳號狀態(false停用 true正常)",
                    "type": "boolean"
                },
                "updateTime": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "userType": {
                    "description": "是否為系統用戶",
                    "type": "boolean"
                }
            }
        },
        "dto.PageDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "searchCategory": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                },
                "sortColumn": {
                    "type": "string"
                }
            }
        },
        "dto.PageForMultSearchDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "search": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "sortColumn": {
                    "type": "string"
                }
            }
        },
...
        "forestagedto.ProductionDTO": {
            "type": "object",
            "properties": {
                "pageData": {
                    "$ref": "#/definitions/dto.PageDTO"
                },
                "productionList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forestagedto.ProductionData"
                    }
                }
            }
        },
        "forestagedto.ProductionData": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "images": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "priceMin": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "forestagedto.ProductionDetailDTO": {
            "type": "object",
            "properties": {
                "production": {
                    "$ref": "#/definitions/forestagedto.ProductionDetailData"
                }
            }
        },
        "forestagedto.ProductionDetailData": {
            "type": "object",
            "properties": {
                "attribute": {
                    "type": "string"
                },
                "categories": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "historicalSold": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "images": {
                    "type": "string"
                },
                "likedCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "priceMin": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}